+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add systemd unit and script to compute fairness values for k8s API server in controlplane.
+- Add internal.advancedConfiguration.kubelet to configure system and k8s reserved resources. 
 
 ## [0.3.0] - 2024-01-22